chapter_001/src/main/java/ru/job4j/array/Check.java
package ru.job4j.array;

/**
 * @author Medoev Ruslan (mr.r.m3@icloud.com)
 * @version $Id$
 * @since 0.1
 */

public class Check {
    /**
     * Метод проверяет элементы массива на идентичность.
     * @param data Массив.
     * @return true or false.
     */
    public boolean mono(boolean[] data) {
        boolean result = true;
        for (int i = 0; i < data.length - 1; i++) {
            if (data[i] != data[i + 1]) {
                result = false;
                break;
            }
        }
        return result;
    }
}
